% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_cfa_isc.R
\name{build.model}
\alias{build.model}
\title{Create lavaan syntax for various model types}
\usage{
build.model(..., type = c("1f", "hierarchical", "oblique", "bifactor"))
}
\description{
Create lavaan syntax for various model types
}

