% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_cfa_isc.R
\name{fit.isc}
\alias{fit.isc}
\title{Fit confirmatory factor analysis model with two groups}
\usage{
fit.isc(df, f.names = c("f1", "f2"))
}
\arguments{
\item{df}{data.frame with names indicating group membership}

\item{f.names}{a 2-item vector used to grep the 1st and 2nd factor columns}
}
\description{
Fit confirmatory factor analysis model with two groups
}
\examples{
d = data.frame(gen.cordata(.5, 6, 100))
names(d) <- paste0(c(rep(c('f1', 'f2'), each=3)), 1:6)
fit.isc(d)
}

