% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gendata.R
\name{gen.cordata}
\alias{gen.cordata}
\title{Use spectral decomposition to create data with exact correlation matrix}
\usage{
gen.cordata(lambda, nsubs, npoints)
}
\arguments{
\item{lambda}{loadings (correlation with be lambda^2)}

\item{nsubs}{number of participants}

\item{npoints}{number of observations per participant}
}
\value{
data.frame with subs columns and npoints rows
}
\description{
Use spectral decomposition to create data with exact correlation matrix
}

