% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gendata.R
\name{gen.corrmat}
\alias{gen.corrmat}
\title{Return correlation matrix using latent variable model}
\usage{
gen.corrmat(g1, g2, lam1, lam2, rho)
}
\arguments{
\item{g1}{indexes for columns of corrmat-to-between. EG g1 = 1:4, g2 = 5:8}

\item{g2}{see g1}

\item{lam1}{loadings of g1 columns on first latent variable}

\item{lam2}{see lam1}

\item{rho}{correlation between latent variables}
}
\description{
Return correlation matrix using latent variable model
}

