% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{gen.meas}
\alias{gen.meas}
\title{Simulate participant data from latent timecourse.}
\usage{
gen.meas(N_part, N_time, var_rat, latent = NULL)
}
\arguments{
\item{N_part}{number of participants}

\item{N_time}{number of timepoints for each participants}

\item{var_rat}{signal variance to noise ratio}

\item{optional}{latent timecourse to draw participants from}
}
\value{
list with two entries. latent is the latent timecourse. measured are the outputs.
}
\description{
Simulate participant data from latent timecourse.
}

