% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classic_isc.R
\name{group.cor}
\alias{group.cor}
\title{Estimate the similarity between two groups}
\usage{
group.cor(M1, M2)
}
\arguments{
\item{M1}{First group. May be data.frame, matrix, or lavaan fit object.}

\item{M2}{Second group. (uneccessary if using lavaan object)}
}
\description{
Estimate the similarity between two groups
}

