% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classic_isc.R
\name{group.rel}
\alias{group.rel}
\title{Estimate the reliability for latent factors underlying items
Note that if it is a lavaan object, it will return a ___ with reliability
for each factor. If it is a data.frame, it will assume a single factor model,
and approximate reliability from item-total correlations.}
\usage{
group.rel(M1)
}
\description{
Estimate the reliability for latent factors underlying items
Note that if it is a lavaan object, it will return a ___ with reliability
for each factor. If it is a data.frame, it will assume a single factor model,
and approximate reliability from item-total correlations.
}

