% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classic_isc.R
\name{item.ic}
\alias{item.ic}
\title{Calculate inter-item correlation (internal consistency).
Note that this returns all of the relevant correlations (then you can average them).}
\usage{
item.ic(M, M2 = NULL, lower = TRUE)
}
\arguments{
\item{M}{data.frame or matrix}

\item{M2}{optional data.frame or matrix to cross-correlate}
}
\value{
array of all inter-item correlations
}
\description{
Calculate inter-item correlation (internal consistency).
Note that this returns all of the relevant correlations (then you can average them).
}

