% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classic.R
\name{item.ttlcor}
\alias{item.ttlcor}
\title{Calculate item-total correlation from full data.
Note that this calculates the correlation of each column against the others.}
\usage{
item.ttlcor(M, is.corr = FALSE)
}
\arguments{
\item{M}{data.frame or matrix}

\item{is.corr}{boolean indicating if M is a corraltion matrix}
}
\value{
array with item-total correlation for each column
}
\description{
Calculate item-total correlation from full data.
Note that this calculates the correlation of each column against the others.
}

