% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{lat.cor}
\alias{lat.cor}
\title{Expected correlation between mean timecourses for two perfectly
correlated latent variables.}
\usage{
lat.cor(eta1, eta2, N)
}
\arguments{
\item{eta1}{proportion of variance that is true signal}

\item{eta2}{same but for second latent variable}

\item{N}{number of measures taken for each latent variable}
}
\description{
Expected correlation between mean timecourses for two perfectly
correlated latent variables.
}

